+++ examples/direct.ts
@@ -1,7 +1,7 @@
 import { fetch } from '../src';
 
 const req = await fetch('https://www.example.com', {
-    parseResponse: false
+  parseResponse: false,
 });
 
 console.log(req);