+++ tests/body.test.ts
@@ -1,0 +1,54 @@
+import { expect, test } from 'bun:test';
+import { fetch } from '../src';
+
+test('body', async () => {
+  const formData = new FormData();
+  // Create a Blob to simulate a file.
+  const fileContent = 'Hello, world!';
+  const blob = new Blob([fileContent], { type: 'text/plain' });
+  formData.append('test_file', blob, 'hello.txt');
+
+  const test_formData = await fetch<{ files: Record<'test_file', string> }>(
+    'https://httpbin.org/anything',
+    {
+      body: formData,
+    }
+  );
+
+  const stream = new ReadableStream({
+    start(controller) {
+      controller.enqueue(new TextEncoder().encode('Hello, '));
+      controller.enqueue(new TextEncoder().encode('world!'));
+      controller.close();
+    },
+  });
+
+  const test_stream = await fetch<{ data: string }>(
+    'https://httpbin.org/anything',
+    {
+      method: 'POST',
+      body: stream,
+      headers: {
+        'content-type': 'text/plain;charset=utf-8',
+      },
+    }
+  );
+
+  const params = new URLSearchParams({ foo: 'bar', baz: 'qux' });
+
+  const test_params = await fetch<{ form: Record<string, any> }>(
+    'https://httpbin.org/anything',
+    {
+      method: 'POST',
+      body: params,
+    }
+  );
+
+  expect(JSON.stringify(test_params.response.form)).toBe(
+    JSON.stringify({ baz: 'qux', foo: 'bar' })
+  ); // httpbin returns form in incorrect order for some reason
+
+  expect(test_stream.response.data).toBe('Hello, world!');
+
+  expect(test_formData.response.files.test_file).toBe(fileContent);
+});